updating for sample age being double

+++ src/main/java/edu/mcw/rgd/expression/SampleEditController.java
@@ -61,9 +61,9 @@ public class SampleEditController implements Controller {
                     s.setLifeStage(request.getParameter("lifeStage" + i));
                     s.setNotes(request.getParameter("notes"+i));
                     if (request.getParameter("ageHigh" + i) != null && !request.getParameter("ageHigh" + i).isEmpty())
-                        s.setAgeDaysFromHighBound(Integer.parseInt(request.getParameter("ageHigh" + i)));
+                        s.setAgeDaysFromHighBound(Double.parseDouble(request.getParameter("ageHigh" + i)));
                     if (request.getParameter("ageLow" + i) != null && !request.getParameter("ageLow" + i).isEmpty() )
-                        s.setAgeDaysFromLowBound(Integer.parseInt(request.getParameter("ageLow" + i)));
+                        s.setAgeDaysFromLowBound(Double.parseDouble(request.getParameter("ageLow" + i)));
 
                     pdao.updateSample(s);
                 }
